requirements.txt
asgiref==3.11.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
crispy-bootstrap5==2025.6
cryptography==46.0.3
dj-database-url==3.0.1
Django==5.2.8
django-allauth==65.13.1
django-crispy-forms==2.5
gunicorn==23.0.0
idna==3.11
packaging==25.0
psycopg2-binary==2.9.11
pycparser==2.23
PyJWT==2.10.1
python-dotenv==1.2.1
requests==2.32.5
sqlparse==0.5.4
tzdata==2025.2
urllib3==2.6.0
whitenoise==6.11.0
